str0m

Module change

Ways to change the Rtc session. SDP or Direct.

str0m has two main APIs for changing the WebRTC session.

SDP API. The common way to talk to browsers using SDP OFFER/ANSWER negotiations. Rtc::sdp_api()

Direct API. Makes changes directly to the session without any negotiation. Rtc::direct_api()

Direct API

The direct API is a lower level API which typically can’t be mixed with the SDP API. If you make changes directly to the session, the remote side would not be aware of them unless you construct some “other way” keeping the two peers in sync.

Structs

DirectApi

Direct change strategy.

DtlsCert

Certificate used for DTLS.

Fingerprint

Certificate fingerprint.

SdpAnswer

SDP answer. Answers can be serialized via serde.

SdpApi

Changes to the Rtc via SDP Offer/Answer dance.

SdpOffer

SDP offer. Offers can be serialized via serde.

SdpPendingOffer

Pending offer from a previous Rtc::sdp_api() call.