str0m::media

Struct MediaTime

pub struct MediaTime(/* private fields */);

Media timeline offset represented by a count (numerator) / frequency (denominator) in seconds.

The numerator is typically the packet time of an Rtp header. The denominator is the clock frequency of the media source (typically 90kHz for video and 48kHz for audio). The denominator is guaranteed to be a positive integer while the numerator could be positive, negative, or zero.

The frequency can be found in the negotiated payload parameters for a media writer.

// Obtain mid from Event::MediaAdded
let mid: Mid = todo!();

// Create a media writer for the mid.
let writer = rtc.writer(mid).unwrap();

// Get the payload type (pt) for the wanted codec.
let params = writer.payload_params().nth(0).unwrap();

// Obtain the frequency for the selected codec.
let freq = params.spec().clock_rate;

let mtime = MediaTime::new(2000, freq);

Implementations

impl MediaTime

pub const ZERO: MediaTime

The additive identity: 0/1.

pub const fn new(numer: i64, denom: Frequency) -> Self

Construct a MediaTime from a guaranteed non-zero Frequency.

pub const fn numer(&self) -> i64

The numerator of the offset time.

pub const fn denom(&self) -> u32

The denominator of the offset time.

pub const fn frequency(&self) -> Frequency

The Frequency of the offset time.

pub const fn from_micros(v: i64) -> MediaTime

Convenience constructor for numbers of microseconds (v/1_000_000).

pub const fn from_fixed_point_6_18(v: i64) -> MediaTime

Convenience constructor for numbers of 24-bit 6.18 fixed point units (v/2^18).

pub const fn from_90khz(v: i64) -> MediaTime

Convenience constructor for numbers of 90kHz units (v/90_000).

pub const fn from_millis(v: i64) -> MediaTime

Convenience constructor for numbers of milliseconds (v/1000).

pub const fn from_hundredths(v: i64) -> MediaTime

Convenience constructor for numbers of hundredths of seconds (v/100).

pub const fn from_secs(v: i64) -> MediaTime

Convenience constructor for numbers of seconds (v/1).

pub fn from_seconds(v: impl Into<f64>) -> MediaTime

Convenience constructor for floating point fractions of seconds as microsecond units.

pub fn as_seconds(&self) -> f64

A floating point fraction second representation.

pub const fn as_micros(&self) -> i64

A microsecond representation.

pub const fn is_zero(&self) -> bool

Predicate for checking that the numerator is 0.

pub const fn abs(self) -> MediaTime

The absolute value of the offset time.

pub const fn rebase(self, denom: Frequency) -> MediaTime

Convert this offset time to have a different denominator (frequency). This conversion may lose precision and, after arithmetic operations with other times of higher frequencies, may have a higher frequency.