str0m::rtp

Module vla

Video Layers Allocation RTP Header Extension

Structs

ResolutionAndFramerate

A resolution and a frame rate, tied together because that’s how it’s formed in the header extension. Either they are both there or neither are there.

Serializer

Serializer of the Video Layers Allocation Header Extension

SimulcastStreamAllocation

An allocation for a simulcast stream, which may contain up to 4 allocations for spatial layers. There may be up to 5 of these per top-level allocation.

SpatialLayerAllocation

An allocation for a spatial layer, which may contain up to 5 allocations for temporal layers. There may be up to 4 per simulcast stream. Also contains an optional resolution and framerate.

TemporalLayerAllocation

An allocation for a temporal layer. There may be up to 5 per spatial layer.

VideoLayersAllocation

Top-level “allocation” for the Video Layers Allocation RTP Header Extension Contains allocations for many simulcast streams, which contain many spatial layers. In practice, there are either many simulcast streams with 1 spatial layer each (simulcast) or 1 simulcast stream with many spatial layers (SVC) or 1 simulcast stream with 1 spatial layer (only temporal layers used). But theoretically, you could have 5 simulcast streams with 4 spatial layers each.

Constants

URI

URI for the Video Layers Allocation RTP Header Extension