str0m

Struct Rtc

pub struct Rtc { /* private fields */ }

Instance that does WebRTC. Main struct of the entire library.

Usage

let mut rtc = Rtc::new();

loop {
    let timeout = match rtc.poll_output().unwrap() {
        Output::Timeout(v) => v,
        Output::Transmit(t) => {
            // TODO: Send data to remote peer.
            continue; // poll again
        }
        Output::Event(e) => {
            // TODO: Handle event.
            continue; // poll again
        }
    };

    // TODO: Wait for one of two events, reaching `timeout`
    //       or receiving network input. Both are encapsulated
    //       in the Input enum.
    let input: Input = todo!();

    rtc.handle_input(input).unwrap();
}

Implementations

impl Rtc

pub fn new() -> Self

Creates a new instance with default settings.

To configure the instance, use RtcConfig.

use str0m::Rtc;

let rtc = Rtc::new();

pub fn builder() -> RtcConfig

Creates a config builder that configures an Rtc instance.

let rtc = Rtc::builder()
    .set_ice_lite(true)
    .build();

pub fn is_alive(&self) -> bool

Tests if this instance is still working.

Certain events will straight away disconnect the Rtc instance, such as the DTLS fingerprint from the setup not matching that of the TLS negotiation (since that would potentially indicate a MITM attack!).

The instance can be manually disconnected using Rtc::disconnect().

let mut rtc = Rtc::new();

assert!(rtc.is_alive());

rtc.disconnect();
assert!(!rtc.is_alive());

pub fn disconnect(&mut self)

Force disconnects the instance making Rtc::is_alive() return false.

This makes Rtc::poll_output and Rtc::handle_input go inert and not produce anymore network output or events.

let mut rtc = Rtc::new();

rtc.disconnect();
assert!(!rtc.is_alive());

pub fn add_local_candidate(&mut self, c: Candidate)

Add a local ICE candidate. Local candidates are socket addresses the Rtc instance use for communicating with the peer.

This library has no built-in discovery of local network addresses on the host or NATed addresses via a STUN server or TURN server. The user of the library is expected to add new local candidates as they are discovered.

In WebRTC lingo, the Rtc instance is permanently in a mode of Trickle Ice. It’s however advisable to add at least one local candidate before starting the instance.

let mut rtc = Rtc::new();

let a = "127.0.0.1:5000".parse().unwrap();
let c = Candidate::host(a, "udp").unwrap();

rtc.add_local_candidate(c);

pub fn add_remote_candidate(&mut self, c: Candidate)

Add a remote ICE candidate. Remote candidates are addresses of the peer.

For SdpApi: Remote candidates are typically added via receiving a remote SdpOffer or SdpAnswer.

However for the case of Trickle Ice, this is the way to add remote candidates that are “trickled” from the other side.

let mut rtc = Rtc::new();

let a = "1.2.3.4:5000".parse().unwrap();
let c = Candidate::host(a, "udp").unwrap();

rtc.add_remote_candidate(c);

pub fn is_connected(&self) -> bool

Checks if we are connected.

This tests both if we have ICE connection and DTLS is ready.

pub fn sdp_api(&mut self) -> SdpApi<'_>

Make changes to the Rtc session via SDP.

let mut rtc = Rtc::new();

let mut changes = rtc.sdp_api();
let mid_audio = changes.add_media(MediaKind::Audio, Direction::SendOnly, None, None);
let mid_video = changes.add_media(MediaKind::Video, Direction::SendOnly, None, None);

let (offer, pending) = changes.apply().unwrap();
let json = serde_json::to_vec(&offer).unwrap();

// Send json OFFER to remote peer. Receive an answer back.
let answer: SdpAnswer = todo!();

rtc.sdp_api().accept_answer(pending, answer).unwrap();

pub fn direct_api(&mut self) -> DirectApi<'_>

Makes direct changes to the Rtc session.

This is a low level API. For “normal” use via SDP, see Rtc::sdp_api().

pub fn writer(&mut self, mid: Mid) -> Option<Writer<'_>>

Send outgoing media data (samples) or request keyframes.

Returns None if the direction isn’t sending (sendrecv or sendonly).

let mut rtc = Rtc::new();

// add candidates, do SDP negotiation
let mid: Mid = todo!(); // obtain mid from Event::MediaAdded.

// Writer for this mid.
let writer = rtc.writer(mid).unwrap();

// Get incoming media data from another peer
let data: MediaData = todo!();

// Match incoming PT to an outgoing PT.
let pt = writer.match_params(data.params).unwrap();

writer.write(pt, data.network_time, data.time, data.data).unwrap();

This is a sample level API: For RTP level see DirectApi::stream_tx() and DirectApi::stream_rx().

pub fn media(&self, mid: Mid) -> Option<&Media>

Currently configured media.

Read only access. Changes are made via Rtc::sdp_api() or Rtc::direct_api().

pub fn poll_output(&mut self) -> Result<Output, RtcError>

Poll the Rtc instance for output. Output can be three things, something to Transmit via a UDP socket (maybe via a TURN server). An Event, such as receiving media data, or a Timeout.

The user of the library is expected to continuously call this function and deal with the output until it encounters an Output::Timeout at which point no further output is produced (if polled again, it will result in just another timeout).

After exhausting the poll_output, the function will only produce more output again when one of two things happen:

The polled timeout is reached.

New network input.

See Rtc instance documentation for how this is expected to be used in a loop.

pub fn accepts(&self, input: &Input<'_>) -> bool

Check if this Rtc instance accepts the given input. This is used for demultiplexing several Rtc instances over the same UDP server socket.

Input::Timeout is always accepted. Input::Receive is tested against the nominated ICE candidate. If that doesn’t match and the incoming data is a STUN packet, the accept call is delegated to the ICE agent which recognizes the remote peer from a=ufrag/a=password credentials negotiated in the SDP. If that also doesn’t match, all remote ICE candidates are checked for a match.

In a server setup, the server would try to find an Rtc instances using Rtc::accepts(). The first found instance would be given the input via Rtc::handle_input().

// A vec holding the managed rtc instances. One instance per remote peer.
let mut rtcs = vec![Rtc::new(), Rtc::new(), Rtc::new()];

// Configure instances with local ice candidates etc.

loop {
    // TODO poll_timeout() and handle the output.

    let input: Input = todo!(); // read network data from socket.
    for rtc in &mut rtcs {
        if rtc.accepts(&input) {
            rtc.handle_input(input).unwrap();
        }
    }
}

pub fn handle_input(&mut self, input: Input<'_>) -> Result<(), RtcError>

Provide input to this Rtc instance. Input is either a Input::Timeout for some time that was previously obtained from Rtc::poll_output(), or Input::Receive for network data.

Both the timeout and the network data contains a std::time::Instant which drives time forward in the instance. For network data, the intention is to record the time of receiving the network data as precise as possible. This time is used to calculate things like jitter and bandwidth.

It’s always okay to call Rtc::handle_input() with a timeout, also before the time obtained via Rtc::poll_output().

let mut rtc = Rtc::new();

loop {
    let timeout: Instant = todo!(); // rtc.poll_output() until we get a timeout.

    let input: Input = todo!(); // wait for network data or timeout.
    rtc.handle_input(input);
}

pub fn channel(&mut self, id: ChannelId) -> Option<Channel<'_>>

Obtain handle for writing to a data channel.

This is first available when a ChannelId is advertised via Event::ChannelOpen. The function returns None also for IDs from SdpApi::add_channel().

Incoming channel data is via the Event::ChannelData event.

let mut rtc = Rtc::new();

let cid: ChannelId = todo!(); // obtain channel id from Event::ChannelOpen
let channel = rtc.channel(cid).unwrap();
// TODO write data channel data.

pub fn bwe(&mut self) -> Bwe<'_>

Configure the Bandwidth Estimate (BWE) subsystem.

Only relevant if BWE was enabled in the RtcConfig::enable_bwe()

pub fn codec_config(&self) -> &CodecConfig

The codec configs for sending/receiving data.

The configurations can be set with RtcConfig before setting up the session, and they might be further updated by SDP negotiation.